Create AmazonScrapper.php
* AmazonScrapper v1.0 (07/2014)
This is a simple Scrapper Engine for Amazon WebSite
It basically searches for important stuff inside it, 
as Categories, Sub-Categories and Products details.
THIS CODE IS UNDER GPL V3 LICENSE

Any suggestions? Write dineu.assis@gmail.com

This file works as a Controller, and also has a sample at the end, where it retrieves all Amazon Categories as a linked list.